requirements.txt
better-ffmpeg-progress
pycryptodomex
requests
python-dateutil
pypresence
m3u8
setuptools
beautifulsoup4
tqdm
moviepy
pyyaml